Add files via upload
Script for one of the enemy types in Project Anarchy. Enemy switches from 3 different states, follow, attack and death.
Follow state uses Unity's Nav Mesh System to guide the enemy towards the player.
Attack state is triggered by a proximity check, pausing the position of the enemy and playing an animation and dealing damage if the player is within a certain range triggered by an animation event.
Death state is when the enemy dies from being shot by the player, disabling the Animator component and using ragdoll with a dissolve effect to destroy the enemy gameObject.